+++ src/main/java/org/broadinstitute/dsm/DSMServer.java
@@ -35,7 +35,6 @@ import org.broadinstitute.dsm.careevolve.Provider;
 import org.broadinstitute.dsm.jetty.JettyConfig;
 import org.broadinstitute.dsm.jobs.*;
 import org.broadinstitute.dsm.log.SlackAppender;
-import org.broadinstitute.dsm.pubsub.ElasticExportSubscription;
 import org.broadinstitute.dsm.pubsub.PubSubResultMessageSubscription;
 import org.broadinstitute.dsm.pubsub.DSMtasksSubscription;
 import org.broadinstitute.dsm.route.*;
@@ -106,7 +105,6 @@ public class DSMServer extends BasicServer {
     public static final String GCP_PATH_TO_PUBSUB_SUB = "pubsub.subscription";
     public static final String GCP_PATH_TO_DSS_TO_DSM_SUB = "pubsub.dss_to_dsm_subscription";
     public static final String GCP_PATH_TO_DSM_TO_DSS_TOPIC = "pubsub.dsm_to_dss_topic";
-    public static final String GCP_PATH_TO_ELASTIC_EXPORT_SUB = "pubsub.elastic_export_subscription";
     public static final String GCP_PATH_TO_DSM_TASKS_SUB = "pubsub.dsm_tasks_subscription";
 
     private static Map<String, JsonElement> ddpConfigurationLookup = new HashMap<>();
@@ -325,7 +323,6 @@ public class DSMServer extends BasicServer {
         String projectId = cfg.getString(GCP_PATH_TO_PUBSUB_PROJECT_ID);
         String subscriptionId = cfg.getString(GCP_PATH_TO_PUBSUB_SUB);
         String dsmToDssSubscriptionId = cfg.getString(GCP_PATH_TO_DSS_TO_DSM_SUB);
-        String elasticExportSubscriptionId = cfg.getString(GCP_PATH_TO_ELASTIC_EXPORT_SUB);
         String DSMtasksSubscriptionId = cfg.getString(GCP_PATH_TO_DSM_TASKS_SUB);
 
         logger.info("Setting up pubsub for {}/{}", projectId, subscriptionId);
@@ -374,12 +371,6 @@ public class DSMServer extends BasicServer {
 
         try {
             PubSubResultMessageSubscription.dssToDsmSubscriber(projectId, dsmToDssSubscriptionId);
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-
-        try {
-            ElasticExportSubscription.subscribeElasticExport(projectId, elasticExportSubscriptionId);
         } catch (Exception e) {
             e.printStackTrace();
         }
