--- src/main/java/org/broadinstitute/dsm/pubsub/ElasticExportSubscription.java
@@ -1,64 +1,0 @@
-package org.broadinstitute.dsm.pubsub;
-
-import com.google.api.gax.core.ExecutorProvider;
-import com.google.api.gax.core.InstantiatingExecutorProvider;
-import com.google.cloud.pubsub.v1.AckReplyConsumer;
-import com.google.cloud.pubsub.v1.MessageReceiver;
-import com.google.cloud.pubsub.v1.Subscriber;
-import com.google.gson.Gson;
-import com.google.pubsub.v1.ProjectSubscriptionName;
-import com.google.pubsub.v1.PubsubMessage;
-import org.broadinstitute.dsm.export.ExportToES;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import java.util.concurrent.TimeUnit;
-import java.util.concurrent.TimeoutException;
-
-public class ElasticExportSubscription {
-
-    private static final Logger logger = LoggerFactory.getLogger(ElasticExportSubscription.class);
-    private static final Gson gson = new Gson();
-
-    public static void subscribeElasticExport(String projectId, String subscriptionId) {
-        // Instantiate an asynchronous message receiver.
-        MessageReceiver receiver =
-                (PubsubMessage message, AckReplyConsumer consumer) -> {
-                    // Handle incoming message, then ack the received message.
-                    logger.info("Got message with Id: " + message.getMessageId());
-                    consumer.ack();
-                    String data = message.getData() != null ? message.getData().toStringUtf8() : null;
-                    ExportPayload payload = gson.fromJson(data, ExportPayload.class);
-                    ExportToES.exportObjectsToES(payload);
-                };
-
-        Subscriber subscriber = null;
-        ProjectSubscriptionName resultSubName = ProjectSubscriptionName.of(projectId, subscriptionId);
-        ExecutorProvider resultsSubExecProvider = InstantiatingExecutorProvider.newBuilder().setExecutorThreadCount(1).build();
-        subscriber = Subscriber.newBuilder(resultSubName, receiver)
-                .setParallelPullCount(1)
-                .setExecutorProvider(resultsSubExecProvider)
-                .setMaxAckExtensionPeriod(org.threeten.bp.Duration.ofSeconds(120))
-                .build();
-        try {
-            subscriber.startAsync().awaitRunning(1L, TimeUnit.MINUTES);
-            logger.info("Started pubsub subscription receiver for dsm elastic export");
-        }
-        catch (TimeoutException e) {
-            throw new RuntimeException("Timed out while starting pubsub subscription or dsm elastic export", e);
-        }
-    }
-
-    public static class ExportPayload {
-        private String index;
-        private String study;
-
-        public String getIndex() {
-            return index;
-        }
-
-        public String getStudy() {
-            return study;
-        }
-    }
-}
