+++ src/test/java/org/broadinstitute/dsm/pubsub/UpdateWorkflowStatusTest.java
@@ -61,7 +61,7 @@ public class UpdateWorkflowStatusTest {
     public void testUpdateProbandStatusInDB() {
         String workflow = "REGISTRATION_STATUS";
         String status = "ENROLLED";
-        DSMtasksSubscription.updateProbandStatusInDB(workflow, status, participantDataDto, DSMtasksSubscription.RGP);
+        WorkflowStatusUpdate.updateProbandStatusInDB(workflow, status, participantDataDto, WorkflowStatusUpdate.RGP);
         String data = participantDataDao.get(participantDataId).orElseThrow().getData();
         JsonObject dataJsonObject = gson.fromJson(data, JsonObject.class);
         Assert.assertEquals(status, dataJsonObject.get(workflow).getAsString());
